requirements.txt
# Core ML/DL
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Data
numpy>=1.26.0
pandas>=2.2.0
yfinance>=0.2.36
pykrx>=1.0.45

# Config
pyyaml>=6.0.1
omegaconf>=2.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# RL
gymnasium>=0.29.0
stable-baselines3>=2.2.0

# Scikit-learn (for metrics, preprocessing)
scikit-learn>=1.4.0

# Logging & monitoring
tensorboard>=2.16.0
tqdm>=4.66.0
loguru>=0.7.2

# Storage
h5py>=3.10.0
pyarrow>=15.0.0

# Statistics
scipy>=1.12.0
statsmodels>=0.14.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
